example refactor
