+++ src/generate/generators/ExtensionIncre.ts
@@ -40,7 +40,7 @@ export class AzExtensionIncrementalGenerator extends GeneratorBase {
     }
 
     public async generateAll(): Promise<void> {
-        const { extensionHandler, configHandler } = this.model.GetHandler();
+        const { extensionHandler, configHandler, exampleHandler } = this.model.GetHandler();
         this.files[
             path.join(this.azDirectory, PathConstants.generatedFolder, PathConstants.paramsFile)
         ] = GenerateAzureCliParams(this.model, this.isDebugMode);
@@ -116,14 +116,14 @@ export class AzExtensionIncrementalGenerator extends GeneratorBase {
 
         await this.generateIncrementalSingleAndAddtoOutput(new CliTestInit(this.model));
         await this.generateIncrementalSingleAndAddtoOutput(new CliTestStep(this.model), true);
-        for (const testGroup of extensionHandler.Extension_TestScenario
-            ? Object.getOwnPropertyNames(extensionHandler.Extension_TestScenario)
+        for (const testGroup of exampleHandler.Example_TestScenario
+            ? Object.getOwnPropertyNames(exampleHandler.Example_TestScenario)
             : []) {
             await this.generateIncrementalSingleAndAddtoOutput(
                 new CliTestScenario(
                     this.model,
                     PathConstants.incTestScenarioFile(testGroup),
-                    extensionHandler.Extension_TestScenario[testGroup],
+                    exampleHandler.Example_TestScenario[testGroup],
                     testGroup,
                 ),
                 true,
@@ -136,7 +136,7 @@ export class AzExtensionIncrementalGenerator extends GeneratorBase {
                 true,
             );
         }
-        this.model
+        exampleHandler
             .GetResourcePool()
             .generateArmTemplate(
                 this.files,
