+++ src/generate/renders/CliMeta.ts
@@ -7,8 +7,8 @@ import { CodeModelAz } from '../codemodel/CodeModelAz';
 import { putToZip } from '../../utils/inplace';
 
 export function GenerateMetaFile(model: CodeModelAz) {
-    const { configHandler } = model.GetHandler();
-    const output: string[] = JSON.stringify(model.GetMetaData(), null, 2).split('\n');
+    const { configHandler, exampleHandler } = model.GetHandler();
+    const output: string[] = JSON.stringify(exampleHandler.GetMetaData(), null, 2).split('\n');
 
     putToZip(configHandler.azOutputFolder, 'meta.txt', output);
 }
