+++ src/generate/renders/tests/CliTestScenario.ts
@@ -3,12 +3,13 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *-------------------------------------------------------------------------------------------- */
 import * as path from 'path';
-import { CodeModelAz, CommandExample } from '../../codemodel/CodeModelAz';
+import { CodeModelAz } from '../../codemodel/CodeModelAz';
 import { CliTestStep } from './CliTestStep';
 import { ToMultiLine, Capitalize } from '../../../utils/helper';
 import { HeaderGenerator } from '../Header';
 import { TemplateBase } from '../TemplateBase';
 import { CodeGenConstants, PathConstants } from '../../../utils/models';
+import { CommandExample } from '../../codemodel/Example';
 
 export class CliTestScenario extends TemplateBase {
     public configValue: any;
@@ -59,8 +60,8 @@ export class CliTestScenario extends TemplateBase {
     }
 
     private GenerateAzureCliTestScenario(model: CodeModelAz, config: any, scenarioName: string) {
-        const { configHandler } = model.GetHandler();
-        const commandParams = model.GatherInternalResource();
+        const { configHandler, exampleHandler } = model.GetHandler();
+        const commandParams = exampleHandler.GatherInternalResource();
         config.unshift({ function: `setup_${scenarioName}` });
         config.push({ function: `cleanup_${scenarioName}` });
 
@@ -72,7 +73,7 @@ export class CliTestScenario extends TemplateBase {
         const testClassName = Capitalize(this.groupName) + scenarioName + 'Test';
         classInfo.push('class ' + testClassName + '(ScenarioTest):');
 
-        const subscriptionId = model.GetSubscriptionKey();
+        const subscriptionId = exampleHandler.GetSubscriptionKey();
         if (subscriptionId) {
             initiates.push('        self.kwargs.update({');
             initiates.push(`            '${subscriptionId}': self.get_subscription_id()`);
@@ -103,7 +104,7 @@ export class CliTestScenario extends TemplateBase {
         );
 
         function buildSenario(template: CliTestScenario, outputFunc: string[], minimum: boolean) {
-            model.GetResourcePool().clearExampleParams();
+            exampleHandler.GetResourcePool().clearExampleParams();
 
             // go through the examples to generate steps
             for (let ci = 0; ci < config.length; ci++) {
@@ -115,7 +116,7 @@ export class CliTestScenario extends TemplateBase {
                     let found = false;
                     const examples: CommandExample[] = [];
                     let exampleIdx = -1;
-                    for (const exampleCmd of model.FindExampleById(
+                    for (const exampleCmd of exampleHandler.FindExampleById(
                         exampleId,
                         commandParams,
                         examples,
@@ -125,7 +126,7 @@ export class CliTestScenario extends TemplateBase {
                         exampleIdx += 1;
                         if (exampleCmd && exampleCmd.length > 0) {
                             found = true;
-                            const checks = model.GetExampleChecks(examples[exampleIdx]);
+                            const checks = exampleHandler.GetExampleChecks(examples[exampleIdx]);
                             functionName = CliTestStep.ToFunctionName(
                                 { name: examples[exampleIdx].Id },
                                 exampleCmd[0],
@@ -181,9 +182,9 @@ export class CliTestScenario extends TemplateBase {
                         );
                         if (
                             functionName.startsWith('setup_') &&
-                            model.GetResourcePool().hasTestResourceScenario
+                            exampleHandler.GetResourcePool().hasTestResourceScenario
                         ) {
-                            steps.push(...model.GetResourcePool().setupWithArmTemplate());
+                            steps.push(...exampleHandler.GetResourcePool().setupWithArmTemplate());
                         } else {
                             steps.push('    pass');
                         }
