+++ test/unittest/render-getActionRenderData.ts
@@ -42,7 +42,8 @@ describe('getActionsRender', () => {
         await entry.init();
 
         const codeModel = new CodeModelCliImpl(session);
-        codeModel.GenerateTestInit();
+        const { exampleHandler } = codeModel.GetHandler();
+        exampleHandler.GenerateTestInit();
 
         model = codeModel;
     }
