+++ test/unittest/render-getCommandsRenderData.ts
@@ -42,7 +42,8 @@ describe('getCommandsRender', () => {
         await entry.init();
 
         const codeModel = new CodeModelCliImpl(session);
-        codeModel.GenerateTestInit();
+        const { exampleHandler } = codeModel.GetHandler();
+        exampleHandler.GenerateTestInit();
 
         model = codeModel;
     }
